openai_agents_project/pyproject.toml
[project]
name = "openai-agents-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "emoji>=2.14.1",
    "openai>=1.99.1",
    "openai-agents>=0.2.4",
    "python-dotenv>=1.1.1",
]
